ocbf/requirement.txt
git+https://github.com/bigd4/PyNEP.git
maml==2024.6.13
matplotlib==3.8.0
path==16.14.0
ase==3.28.0
spglib==2.7.0
dscribe==2.1.1
Cython==3.0.11
numpy==1.26.4
pyyaml==6.0.2
pip==24.2

